requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
motor==3.7.*
pymongo==4.12.*
redis[hiredis]==5.2.*
pydantic==2.11.*
pydantic-settings==2.9.*
python-dotenv==1.1.*
openai>=1.30,<2
boto3>=1.35,<2
python-multipart>=0.0.9
